Add files via upload
I've found the "box model" a bit confusing when just reading about it, so I've decided to sort of tweak the approach outlined in my course. I decided to try working on certain parts of the webpage at a time, filling stuff in as I get those parts to look the way I want them to. In this version, I just wanted to try to get a basic navbar and footer without really worrying too much about how they looked. The links are empty and the footer just has text so I could make sure I positioned it correctly.