Update index.html
#Changed 
Got rid of the expanding cards thing in the projects section. Cool effect but I couldn't get it to work the way I wanted it to, so it looked kind of tacky.
Changed the size of the cards 
Added a hover "lift" effect 
Added a couple more cards to test responsiveness. Seems to be working (for now...)